feat(cluster-credentials-modal): new UI and STS
