+++ libs/domains/clusters/feature/src/index.ts
@@ -5,6 +5,7 @@ export * from './lib/cluster-action-toolbar/cluster-action-toolbar'
 export * from './lib/cluster-avatar/cluster-avatar'
 export * from './lib/cluster-setup/cluster-setup'
 export * from './lib/cluster-card/cluster-card'
+export * from './lib/cluster-credentials-modal/cluster-credentials-modal'
 export * from './lib/nodepools-resources-settings/nodepools-resources-settings'
 export * from './lib/kubeconfig-preview/kubeconfig-preview'
 export * from './lib/cluster-migration-modal/cluster-migration-modal'
