+++ libs/pages/layout/src/lib/feature/spotlight-trigger/spotlight-trigger.tsx
@@ -11,7 +11,7 @@ export function SpotlightTrigger() {
 
   useEffect(() => {
     const down = (event: KeyboardEvent) => {
-      if (event.key.toLowerCase() === 'k' && (event.metaKey || event.ctrlKey)) {
+      if (event.key?.toLowerCase() === 'k' && (event.metaKey || event.ctrlKey)) {
         event.preventDefault()
         setOpenSpotlight(true)
       }
