+++ libs/shared/console-shared/src/index.ts
@@ -19,7 +19,6 @@ export * from './lib/git-repository-settings/feature/edit-git-repository-setting
 export * from './lib/member-invitation/accept-invitation/feature/accept-invitation-feature'
 export * from './lib/cluster-settings/ui/cluster-general-settings/cluster-general-settings'
 export * from './lib/cluster-settings/feature/cluster-credentials-settings-feature/cluster-credentials-settings-feature'
-export * from './lib/cluster-settings/feature/create-edit-credentials-modal-feature/create-edit-credentials-modal-feature'
 export * from './lib/cluster-settings/ui/card-cluster-feature/card-cluster-feature'
 export * from './lib/add-credit-card-modal/feature/add-credit-card-modal-feature'
 export * from './lib/layout-logs/layout-logs'
