+++ libs/shared/spotlight/feature/src/lib/sub-command/sub-command.tsx
@@ -57,7 +57,7 @@ export function SubCommand({
     // Handler for keyboard shortcuts
     const handleKeyDown = (event: KeyboardEvent) => {
       // Meta+K toggle
-      if (event.key.toLowerCase() === 'k' && (event.metaKey || event.ctrlKey)) {
+      if (event.key?.toLowerCase() === 'k' && (event.metaKey || event.ctrlKey)) {
         event.preventDefault()
         setOpen(!open)
       }
